2023/2/20 kuro2 适应性修改

+++ KuroModifyTool/FileTools.cs
@@ -1,9 +1,13 @@
 ﻿using KuroModifyTool.KuroTable;
 using Microsoft.International.Converters.TraditionalChineseToSimplifiedConverter;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
+using System.Threading;
+using System.Windows.Input;
 
 namespace KuroModifyTool
 {
@@ -27,6 +31,54 @@ namespace KuroModifyTool
             fs.Write(buffer, 0, buffer.Length);
             fs.Flush();
             fs.Close();
+        }
+
+        public static void PackTbl(string srcp, string tarp)
+        {
+            Process p = new Process();
+            //设置要启动的应用程序
+            p.StartInfo.FileName = "cmd.exe";
+            //是否使用操作系统shell启动
+            p.StartInfo.UseShellExecute = false;
+            // 接受来自调用程序的输入信息
+            p.StartInfo.RedirectStandardInput = true;
+            //输出信息
+            p.StartInfo.RedirectStandardOutput = true;
+            // 输出错误
+            p.StartInfo.RedirectStandardError = true;
+            //不显示程序窗口
+            p.StartInfo.CreateNoWindow = true;
+            //启动程序
+            p.StartInfo.Verb = "RunAs";
+
+            p.Start();
+            string strInput = "\"" + Environment.CurrentDirectory + "\\EnComp.exe\" -ec \"" + srcp + "\" \"" + tarp + "\"";
+            //sb.Append(strInput + "\n");
+            p.StandardInput.WriteLine(strInput);
+            p.StandardInput.WriteLine("exit");
+            p.StandardInput.AutoFlush = true;
+            Thread tr = new Thread(new ParameterizedThreadStart(outputT));
+            tr.Start(p);
+            Thread tr1 = new Thread(new ParameterizedThreadStart(outputT));
+            tr1.Start(p);
+            p.WaitForExit();
+            p.Close();
+
+            p = null;
+        }
+
+        private static void outputT(object o)
+        {
+            Process p = (Process)o;
+            string outPut = p.StandardOutput.ReadToEnd();
+            outPut = null;
+        }
+
+        private void errorT(object o)
+        {
+            Process p = (Process)o;
+            string error = p.StandardError.ReadToEnd();
+            error = null;
         }
 
         public static void WriteLog(string text)
@@ -118,6 +170,47 @@ namespace KuroModifyTool
             return skillDic;
         }
 
+        public static Dictionary<byte, string> GetCurrencyDic(string path)
+        {
+            StreamReader sr = new StreamReader(path);
+            Dictionary<byte, string> dic = new Dictionary<byte, string>();
+
+            while (!sr.EndOfStream)
+            {
+                string text = sr.ReadLine();
+
+                if (text == "")
+                {
+                    continue;
+                }
+
+                string[] data = text.Split(':');
+
+                dic.Add(byte.Parse(data[0]), data[1]);
+            }
+            sr.Close();
+
+            return dic;
+        }
+
+        public static T JsonLoad<T>(string path)
+        {
+            StreamReader sr = new StreamReader(path);
+            T t = JsonConvert.DeserializeObject<T>(sr.ReadToEnd());
+
+            sr.Close();
+            return t;
+        }
+
+        public static void JsonSave(object obj, string path)
+        {
+            StreamWriter sw = new StreamWriter(path, false);
+            string json = JsonConvert.SerializeObject(obj);
+
+            sw.Write(json);
+            sw.Close();
+        }
+
         public static bool initConfig(Action action)
         {
             if (!File.Exists(StaticField.ConfigPath))
@@ -135,11 +228,15 @@ namespace KuroModifyTool
                     sw.Close();
                     fs.Close();
 
+                    StaticField.CLEFiles = new List<CLEFile>();
+                    JsonSave(StaticField.CLEFiles, StaticField.CLEFList);
+
                     BakcutFile("t_item.tbl");
                     BakcutFile("t_skill.tbl");
                     BakcutFile("t_shard_skill.tbl");
                     BakcutFile("t_hollowcore.tbl");
-                    BakcutFile("t_artsdriver.tbl");
+                    BakcutFile("t_artsdriver.tbl"); 
+                    BakcutFile("t_voice.tbl");
 
                     return true;
                 }
@@ -153,6 +250,9 @@ namespace KuroModifyTool
                 StreamReader sr = new StreamReader(StaticField.ConfigPath);
 
                 StaticField.GamePath = sr.ReadLine();
+                sr.Close();
+
+                StaticField.CLEFiles = JsonLoad<List<CLEFile>>(".\\KuroList\\CLEFileList.json");
                 return true;
             }
         }
@@ -180,7 +280,14 @@ namespace KuroModifyTool
         {
             if(!File.Exists(".\\bak\\" + name))
             {
-                File.Copy(StaticField.TBLPath + name, ".\\bak\\" + name);
+                if(File.Exists(StaticField.TBLPath + name))
+                {
+                    File.Copy(StaticField.TBLPath + name, ".\\bak\\" + name);
+                }
+                else
+                {
+                    File.Copy(StaticField.TBLPath1 + name, ".\\bak\\" + name);
+                }
             }
         }
